isIdentifierPart should not allow colon or dash
tsgo incorrectly lexes non-ascii identifiers followed by colon as a
single identifier -- in non-TS files only. That is, 'x:' instead of 'x' ':' (where x is
non-ascii). This is incorrect in some, probably most, cases. The place I
noticed it was an object literal.

```ts
f({ 県: '大阪府' });
```

That's because isIdentifierPart allows colon in identifiers for JSX when
languageVariant is provided.
However, in tsc, languageVariant is optional and is only defined in 3
calls from services. All other calls disallow colon, even in JSX files.
In tsgo, languageVariant is always provided, and non-TS files are
treated as JSX.

It's likely that colons and dashes should only be allowed *not* at the
end of an identifier, and that the 3 uses of the check in tsc's services
are also buggy. I haven't read the code to try to construct a test case.

I also haven't checked in a test with this PR. A type or symbol baseline
will easily display it, and the test case I provided above shows it off.