Add toplevel await
1. No incremental reparsing, although the non-incremental reparsing only
reparses the statements that might contain await.
2. No TransformFlags, just an approximation in the parser.

I haven't added all the code to turn off PossibleTopLevelAwait because
I'm not convinced my approach is right. There are 16-17 more places to
save/overwrite the toplevel await tracking bool. I made enough changes
to pass our test suite, that's all.

It's also possible that reparsing, say, a FunctionDeclaration, doesn't
change anything anyway, but it's inefficient to reparse it
when it's not needed.