Fix:allow type annotation w/empty parameters
Previously the empty parameter case was confused with the missing
parameter case because empty slices are optimised to nil slices.

I removed this *only for parseDelimitedList*, but probably should for
other calls to NewSlice as well.

The other option I thought of was to port the isMissingList bool to
NodeList, but that changes the struct from 32 to 40 bytes, which I
believe would be less efficient than a non-nil slice only for
zero-length lists.