Fix Loc of synthetic JSX closing elements
In malformed JSX, some synthetic elements are invented after parsing has
finished and inserted earlier in the tree to attempt to close unclosed
elements. These elements need to specify both a saved pos and saved end,
so I created a new function `finishNodeWithEnd` for that.

I remember also creating this function for parsing JSDoc, which occasionally does
the same thing, so I expect it'll be used more later.