Improve pos for JSDoc
Nodes with JSDoc now have coverage for their entire span. Previously
there were holes because JSDoc nodes didn't include leading whitespace.

I'm not sure this is correct for children of the node with associated
JSDoc. For example, a FunctionDeclaration may span (0, 273] and have
a JSDoc child (0, 38], but its `function` keyword span was parsed as if
JSDoc was whitespace--so it will have a span starting a 0 too, eg (0,
110].