Check JSDoc types
Only improves a few TS tests. I'm not sure how to update the JS tests
now.

The rules for JSVariadicType (`...T`) are simplified: it's a synonym for
`T[]`. Previously it was only `T[]` when used correctly: as a rest
parameter, and was `T | undefined` otherwise.

This makes no sense.

And it turns out that signature help in Strada has always used the
simpler `...T = T[]` rule anyway, so that's what JS users are used to
seeing as long as they don't have checkJs turned on.