Add Host pointer to JSDocTypeExpression,fix name resolution
nameresolver still needs to check for JS nodes because it will often
start a walk upward looking at the *original* type nodes deeply nested
in JSDoc. When it hits the JSDocTypeExpression, it should move over to
the type expression's hosted position in the tree, such as a real
parameter.

The JS code is much less intrusive than before though.