@type tag correctly applies to multiple var decls
Specifically, it applies only to the first. This is how it works in
Strada, but somehow we have no tests of it.

I don't know how toplevel tests are supposed to work. They don't seem to
generate baselines, but in this case all I need is for the test not to
panic to see that it's working.