Add control flow-based inference for this.p in ctor
Note that the baselines *still* show a difference for some auto-typed
results--Strada says `this.p: any` while Corsa says `this.p: string`,
even though both behave as that it's a string when used. I debugged a
test to convince myself that indeed both Strada and Corsa believe the
example property is a string, though.