Clean up GetElementOrPropertyAccessName
Fixes #890

There were two related functions but only one caller needed the .Text()
followup. The new code is stricter and therefore less likely to panic
downstream.

There was also a duplicate implementation of IsModuleExports in checker.
I removed that and cleaned it up as well.

This is also the inauguaral test in cases/conformance/salsa.