Improve destructured `@param` tag matching
Before this PR, `@param`-parameter matching starts by finding the index
of the `@param` tag in the list of tags. This index is used as a
fallback when the tag doesn't match by name. But the index can be wrong
if there are non-`@param` tags earlier in the list of tags, eg a
`@template` tag.

This PR finds the index, only considering `@param` tags.

The current method is still tag-centred, which is not as good as the
parameter-centric method that Strada used--but it's more efficient and
fits better with the other reparser code. The tag-centric method used
here always matches a tag with a parameter if either the name or index
match. Strada matches all tags to all parameters by name first, and only
then matches by index.