Stop getFlowTypeOfProperty overwriting w/base class property
In a bad translation from Typescript, the call to
getTypeOfPropertyInBaseClass always overwrites `initialType`, even when
there's no base class or property in the base class.

I think the reason this only affects JS is that in TS all constructor
control flow properties are auto-typed, so never take this branch.

This subtle bug would have been prevented with strict null checking.