Special-case jsdoc ...T in jsdoc signatures
Currently in Corsa, `...T` is always interpreted as `T[]`, which is considerably
simpler than in Strada. However, there is no
way to express rest parameters in a `@callback` or `@overload` tag. This
PR special-cases `...` in `@param` tags in JSDocSignature, which is how these tags parse,
such that instead of `...T --> T[]`,
it reparses as `...T --> `...paramName: T`.

Notes:
1. This is how it used in unifier, which is a heavily jsdoc'ed
codebase checked with Strada.
2. This PR *removes* the variadic type in reparsing, meaning that the
author needs to provide an array type, like `...[number, number]` for
example. This is still a subset of the Strada semantics.
3.  Code that relies on Strada treating `...` both as rest *and* as
array type in this position gets an error that an array type is
required. This should help other uses update to Corsa semantics.

The new semantics are inconsistent but they are still fairly simple and
should match with how people use `...T` in Strada. See how the new test
based on example code from unifiedjs uses `...T` to mean rest parameter
on `@overload` and to mean `T[]` on a `@param` of the implementation.