This PR allows `@type` tags to apply as a whole to a function. This is a returning Strada feature, but it's no longer implemented as a fallback lookup for a type tag. Instead, `FunctionLikeBase` now has a property `WholeType` that gets set in the reparser. If we can come up with Typescript syntax for this feature, it will work immediately--with a couple of caveats described below.
In the checker, the implementation is basically the same as in Strada, except that it refers to a property on FunctionLikeBase instead of looking up a tag. I put it on FunctionLikeBase because it is more convenient to retrieve, even though it only applies to FunctionLikeWithBodyBase. Notably, there was one check that I skipped because it short-circuits contextual typing in favour of the whole type, which I don't believe is correct--or is perhaps unobservable. When I removed it no tests changed.

I didn't check that declaration emit works, but I think there may be some tests that exercise it. In any case, since the signature's types are pulled from the whole-type when needed, and parameter/return types are supposed to be nil when it's present, I expect that emit will generate types correctly from the signature.

Another difference is that the parameter names of the resulting signature come from the type annotation, not the function declaration. This is consistent with something like
```ts
const x: number = 'xx'
```
where `x: number` when used even though it is initialised with a string, which of course is a string.

Speaking of errors, I improved the wording of the existing error and added another that forbids using `@type` and `@param/@return`. Both errors are worded in terms of JSDoc tags even though we could implement this feature in TS. We can revisit this either after coming up with a good name for this feature, or after we implement it in TS, if ever.

\#\# Open Questions
1. What should this feature be called?
2. Where should the code live?
3. Can the errors be improved further?
4. What would TS syntax for this look like?