finishNode returns its node parameter
This allows it to be used inline, the way it is in Strada. I discussed this with the team weeks ago and we agreed it was a good idea&mdash; and I've had this sitting on my laptop 98% done but I forgot about it till I went on a working vacation and the branch had a bunch of changes.

I didn't do this for withJSDoc, because it currently has a return value that I use to add JSDoc to commonjs nodes in reparseCommonJS. Maybe there's a better way to do that.

I started by asking agent-mode to do this for me, but it currently cannot handle 250 simple changes across 2 files. Before that, I briefly tried letting copilot completions fill it in, but the gaps were often too far apart for NES. So I made a couple of vi macros instead. Now, I would greatly appreciate if agent-mode could make vi macros because it is slow, error-prone and iterative for me.