Fix bogus "no jsdoc types" error in jsdoc
The code dropped the "is in jsdoc" check when it got ported. I noticed this when diffing errors from a native JS code base, three.js I think, in among the new errors that are results of intended cuts.

Notably, this PR shows a quick check for reparsed types: NodeFlagsJSDoc is still set on cloned nodes.
This does not help with non-jsdoc JS structures like expandos and commonjs, though.