fix: make ResponseMessage.Result omitzero
According to the JSON-RPC specification, a response object
must include either a `result` field or an `error` field, but not both.

Without omitzero, if `Result` is nil, it is marshalled as 'null', causing
both `result` and `error` fields to appear in the JSON object when `Error`
is also present, which violates the specification.