fix(parser): panic when parsing JSDoc `@satisfies` with export assignments
The parser was panicking when encountering JSDoc `@satisfies` tags on
module.exports statements. The issue was that KindExportAssignment and
KindJSExportAssignment were incorrectly grouped with nodes that have
an Initializer() method, but they actually use Expression() instead.

Also added missing support for KindCommonJSExport in the Initializer()
and SetInitializer() methods.

Reproduction:
```javascript
/**
 * @satisfies {import('../types.js').SupportVersionTraceMap}
 */
module.exports = {
    zlib: zlib,
    'node:zlib': {
        ...zlib,
        [READ]: { supported: ['14.13.1', '12.20.0'] },
    },
};
```